Sidebar layouts

Create a collapsing sidebar layout by providing a sidebar() object to the sidebar argument of:

layout_sidebar()

Creates a sidebar layout component which can be dropped inside any page() or card() context.

page_navbar(), navs_tab_card(), and navs_pill_card()

Creates a multi page/tab UI with a singular sidebar() (which is shown on every page/tab).

See this article to learn more.

Usage

sidebar(
  ...,
  width = 250,
  position = c("left", "right"),
  open = TRUE,
  id = NULL,
  bg = NULL,
  class = NULL
)

layout_sidebar(
  sidebar = sidebar(),
  ...,
  fillable = FALSE,
  fill = TRUE,
  bg = NULL,
  border = TRUE,
  border_radius = TRUE,
  height = NULL
)

sidebar_open(id, session = get_current_session())

sidebar_close(id, session = get_current_session())

Arguments

...

Unnamed arguments can be any valid child of an htmltools tag and named arguments become HTML attributes on returned UI element. In the case of layout_sidebar(), these arguments are passed to the main content tag (not the sidebar+main content container).

width

A valid CSS unit used for the width of the sidebar.

position

Where the sidebar should appear relative to the main content.

open

Whether or not the sidebar should be open on page load. Provide a value of NA to prevent sidebar from being collapsible.

id

A character string. Required if wanting to re-actively read (or update) the collapsible state in a Shiny app.

bg

A background color. If provided, an accessible contrasting color is provided for the foreground color (consider using a utility class to customize the foreground color).

class

Additional CSS classes for the top-level HTML element.

sidebar

A sidebar() object.

fillable

Whether or not the main content area should be considered a fillable (i.e., flexbox) container.

fill

Whether or not the layout container should grow/shrink to fit a fillable container.

border

Whether or not to add a border.

border_radius

Whether or not to add a border radius.

height

Any valid CSS unit (e.g., height="200px").

session

a shiny session object (the default should almost always be used).

Functions

sidebar_open(): Open a sidebar() (during an active Shiny user session).

sidebar_close(): Close a sidebar() (during an active Shiny user session).