bs_get_variables {bslib}	R Documentation

Retrieve Sass variable values from the current theme

Description

Useful for retriving a variable from the current theme and using the value to inform another R function.

Usage

bs_get_variables(theme, varnames)

bs_get_contrast(theme, varnames)

Arguments

theme
a bs_theme() object.

varnames
a character string referencing a Sass variable in the current theme.

Value

a character string containing a CSS/Sass value. If the variable(s) are not defined, their value is NA.

Examples

vars <- c("body-bg", "body-color", "primary", "border-radius")
bs_get_variables(bs_theme(), varnames = vars)
bs_get_variables(bs_theme(bootswatch = "darkly"), varnames = vars)

bs_get_contrast(bs_theme(), c("primary", "dark", "light"))

library(htmltools)
div(
  class = "bg-primary",
  style = css(
    color = bs_get_contrast(bs_theme(), "primary")
  )
)

[Package bslib version 0.2.4 Index]