Create a Bootstrap theme

Creates a Bootstrap theme object which can be:

Used in any HTML page powered by shiny::bootstrapLib() (e.g., shiny::fluidPage(), shiny::bootstrapPage(), etc).

Used in any output format powered by rmarkdown::html_document() (or rmarkdown::html_document_base()).

Used more generally in any htmltools::tags via bs_theme_dependencies().

These functions (i.e., bs_theme() or bs_theme_update()) allow you to do the following common Bootstrap customization(s):

Choose a (major) Bootstrap version.

Choose a Bootswatch theme (optional).

Customize main colors and fonts via explicitly named arguments (e.g., bg, fg, primary, etc).

Customize other, lower-level, Bootstrap Sass variable defaults via ...

See all Bootstrap 4 variables

See all Bootstrap 3 variables

For less common theming customization(s), you can modify theme objects to:

Add additional Sass/CSS rules (see bs_add_rules() and sass_partial()).

Leverage (new) Sass functions and mixins in those rules (see bs_add_declarations())

These lower-level theming tools build on the concept of a sass::sass_layer(). To learn more, see here.

bs_theme(
  version = version_default(),
  bootswatch = NULL,
  ...,
  bg = NULL,
  fg = NULL,
  primary = NULL,
  secondary = NULL,
  success = NULL,
  info = NULL,
  warning = NULL,
  danger = NULL,
  base_font = NULL,
  code_font = NULL,
  heading_font = NULL
)

bs_theme_update(
  theme,
  ...,
  bg = NULL,
  fg = NULL,
  primary = NULL,
  secondary = NULL,
  success = NULL,
  info = NULL,
  warning = NULL,
  danger = NULL,
  base_font = NULL,
  code_font = NULL,
  heading_font = NULL
)

is_bs_theme(x)

Arguments

version
The major version of Bootstrap to use. A value of '4+3' means Bootstrap 4, but with additional CSS/JS to support BS3 style markup in BS4. Other supported versions include 3 and 4.

bootswatch
The name of a bootswatch theme (see bootswatch_themes() for possible values).

...
arguments passed along to bs_add_variables().

bg
A color string for the background.

fg
A color string for the foreground.

primary
A color to be used for hyperlinks, to indicate primary/default actions, and to show active selection state in some Bootstrap components. Generally a bold, saturated color that contrasts with the theme's base colors.

secondary
A color for components and messages that don't need to stand out. (Not supported in Bootstrap 3.)

success
A color for messages that indicate an operation has succeeded. Typically green.

info
A color for messages that are informative but not critical. Typically a shade of blue-green.

warning
A color for warning messages. Typically yellow.

danger
A color for errors. Typically red.

base_font
The default typeface.

code_font
The typeface to be used for code. Be sure this is monospace!

heading_font
The typeface to be used for heading elements.

theme
a bs_theme() object.

x
an object.

Colors

Colors may be provided in any format that htmltools::parseCssColors() can understand. To control the vast majority of the ('grayscale') color defaults, specify both the fg (foreground) and bg (background) colors. The primary and secondary theme colors are also useful for accenting the main grayscale colors in things like hyperlinks, tabset panels, and buttons.

Fonts

Use base_font, code_font, and heading_font to control the main typefaces. These arguments set new defaults for the relevant font-family CSS properties which does not guarantee the relevant fonts are available in the users system. To ensure the fonts are actually available, use a package like gfonts (if Google Fonts) to download and provide the font files with the HTML site.

Each *_font argument accepts a character vector where each element of that vector can a single unquoted font family name, a single quoted font family name, or a comma-separated list of font families (with individual font family names quoted as necessary). The comma-separated list is useful for specifying "fallback" font families (e.g., generic CSS families like sans-serif or serif) when font(s) may be unavailable.

For example, each example below is valid:

# Single, unquoted
bs_theme(base_font = "Source Sans Pro")
# Single, quoted
bs_theme(base_font = "'Source Sans Pro'")
# Multiple, quoted
bs_theme(base_font = "'Source Sans Pro', sans-serif")
# Combining all of the above
bs_theme(
  base_font = c("Open Sans", "'Source Sans Pro'",
  "'Helvetica Neue', Helvetica, sans-serif")
)

But the following is technically not valid because Source Sans Pro is not quoted (the resulting CSS will contain font-family: Source Sans Pro, sans-serif; which is technically out of the CSS specifications but may still work in some modern browsers).

bs_theme(base_font = "Source Sans Pro, sans-serif")

References

https://getbootstrap.com/docs/4.4/getting-started/theming/

https://rstudio.github.io/sass/

See also

Examples

theme <- bs_theme(
  # Controls the default grayscale palette
  bg = "#202123", fg = "#B8BCC2",
  # Controls the accent (e.g., hyperlink, button, etc) colors
  primary = "#EA80FC", secondary = "#48DAC6",
  base_font = c("Grandstander", "sans-serif"),
  code_font = c("Courier", "monospace"),
  heading_font = "'Helvetica Neue', Helvetica, sans-serif",
  # Can also add lower-level customization
  "input-border-color" = "#EA80FC"
)
if (interactive()) {
  bs_theme_preview(theme)
}

# Lower-level bs_add_*() functions allow you to work more
# directly with the underlying Sass code
theme <- theme %>%
  bs_add_variables("my-class-color" = "red") %>%
  bs_add_rules(".my-class { color: $my-class-color }")