bs_theme_preview {bslib}	R Documentation

Preview the currently set theme

Description

Launches an example shiny app via run_with_themer() and bs_theme_dependencies(). Useful for getting a quick preview of the current theme setting as well as an interactive GUI for tweaking some of the main theme settings.

Usage

bs_theme_preview(theme = bs_theme(), ..., with_themer = TRUE)

Arguments

theme
a bs_theme() object.

...
passed along to shiny::runApp().

with_themer
whether or not to run the app with run_with_themer().

Details

The app that this launches is subject to change.

Value

nothing, this function is called for its side-effects (launching an application).

See Also

run_with_themer()

Examples

theme <- bs_theme(bg = "#6c757d", fg = "white", primary = "orange")
if (interactive()) bs_theme_preview(theme)

[Package bslib version 0.2.4 Index]