nav_prepend {bslib}	R Documentation

Append or prepend nav item to a dropdown menu

Description

Exported for use by shiny::prependTab()/shiny::appendTab(). These functions have been superseded by nav_insert() (i.e., shiny::insertTab()), since it can do everything these functions do (i.e., add a nav() to the start or end of a nav_menu()) and more (i.e., insert a nav() anywhere inside a nav container).

Usage

nav_prepend(
  id,
  nav,
  menu_title,
  select = FALSE,
  session = getDefaultReactiveDomain()
)

nav_append(
  id,
  nav,
  menu_title,
  select = FALSE,
  session = getDefaultReactiveDomain()
)

Arguments

id
a character string used to identify the nav container.

nav
a nav() item.

menu_title
The title of a nav_menu().

select
Should nav be selected upon being inserted?

session
a shiny session object (the default should almost always be used).

[Package bslib version 0.4.1 Index]