Navigation containers

Render a collection of nav() items into a container.

navs_tab(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_pill(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_pill_list(
  ...,
  id = NULL,
  selected = NULL,
  header = NULL,
  footer = NULL,
  well = TRUE,
  fluid = TRUE,
  widths = c(4, 8)
)

navs_hidden(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_bar(
  ...,
  title = NULL,
  id = NULL,
  selected = NULL,
  position = c("static-top", "fixed-top", "fixed-bottom"),
  header = NULL,
  footer = NULL,
  bg = NULL,
  inverse = "auto",
  collapsible = TRUE,
  fluid = TRUE
)

navs_tab_card(..., id = NULL, selected = NULL, header = NULL, footer = NULL)

navs_pill_card(
  ...,
  id = NULL,
  selected = NULL,
  header = NULL,
  footer = NULL,
  placement = c("above", "below")
)

Arguments

...
a collection of nav() items.

id
a character string used for dynamically updating the container (see nav_select()).

selected
a character string matching the value of a particular nav() item to selected by default.

header
UI element(s) (tags) to display above the nav content.

footer
UI element(s) (tags) to display below the nav content.

well
TRUE to place a well (gray rounded rectangle) around the navigation list.

fluid
TRUE to use fluid layout; FALSE to use fixed layout.

widths
Column widths of the navigation list and tabset content areas respectively.

title
The title to display in the navbar

position
Determines whether the navbar should be displayed at the top of the page with normal scrolling behavior ("static-top"), pinned at the top ("fixed-top"), or pinned at the bottom ("fixed-bottom"). Note that using "fixed-top" or "fixed-bottom" will cause the navbar to overlay your body content, unless you add padding, e.g.: tags$style(type="text/css", "body {padding-top: 70px;}")

bg
a CSS color to use for the navbar's background color.

inverse
Either TRUE for a light text color or FALSE for a dark text color. If "auto" (the default), the best contrast to bg is chosen.

collapsible
TRUE to automatically collapse the navigation elements into a menu when the width of the browser is less than 940 pixels (useful for viewing on smaller touchscreen device)

placement
placement of the nav items relative to the content.

See also

Examples

navs <- list(
  nav("A", "a"),
  nav("B", "b")
)

navs_tab(!!!navs)

#> <div class="container-fluid">
#>   <div class="tabbable">
#>     <ul class="nav nav-tabs" data-tabsetid="6824">
#>       <li class="active">
#>         <a href="#tab-6824-1" data-toggle="tab" data-bs-toggle="tab" data-value="A">A</a>
#>       </li>
#>       <li>
#>         <a href="#tab-6824-2" data-toggle="tab" data-bs-toggle="tab" data-value="B">B</a>
#>       </li>
#>     </ul>
#>     <div class="tab-content" data-tabsetid="6824">
#>       <div class="tab-pane active" data-value="A" id="tab-6824-1">a</div>
#>       <div class="tab-pane" data-value="B" id="tab-6824-2">b</div>
#>     </div>
#>   </div>
#> </div>

navs_pill(!!!navs)

#> <div class="container-fluid">
#>   <div class="tabbable">
#>     <ul class="nav nav-pills" data-tabsetid="7325">
#>       <li class="active">
#>         <a href="#tab-7325-1" data-toggle="tab" data-bs-toggle="tab" data-value="A">A</a>
#>       </li>
#>       <li>
#>         <a href="#tab-7325-2" data-toggle="tab" data-bs-toggle="tab" data-value="B">B</a>
#>       </li>
#>     </ul>
#>     <div class="tab-content" data-tabsetid="7325">
#>       <div class="tab-pane active" data-value="A" id="tab-7325-1">a</div>
#>       <div class="tab-pane" data-value="B" id="tab-7325-2">b</div>
#>     </div>
#>   </div>
#> </div>

navs_pill_list(!!!navs)

#> <div class="container-fluid">
#>   <div class="row">
#>     <div class="col-sm-4 well">
#>       <ul class="nav nav-pills nav-stacked" data-tabsetid="4641">
#>         <li class="active">
#>           <a href="#tab-4641-1" data-toggle="tab" data-bs-toggle="tab" data-value="A">A</a>
#>         </li>
#>         <li>
#>           <a href="#tab-4641-2" data-toggle="tab" data-bs-toggle="tab" data-value="B">B</a>
#>         </li>
#>       </ul>
#>     </div>
#>     <div class="col-sm-8">
#>       <div class="tab-content" data-tabsetid="4641">
#>         <div class="tab-pane active" data-value="A" id="tab-4641-1">a</div>
#>         <div class="tab-pane" data-value="B" id="tab-4641-2">b</div>
#>       </div>
#>     </div>
#>   </div>
#> </div>

navs_bar(!!!navs)

#> <nav class="navbar navbar-default navbar-static-top navbar-inverse" role="navigation">
#>   <div class="container-fluid">
#>     <div class="navbar-header">
#>       <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar-collapse-4579" data-bs-toggle="collapse" data-bs-target="#navbar-collapse-4579">
#>         <span class="sr-only">Toggle navigation</span>
#>         <span class="icon-bar"></span>
#>         <span class="icon-bar"></span>
#>         <span class="icon-bar"></span>
#>       </button>
#>       <span class="navbar-brand"></span>
#>     </div>
#>     <div class="navbar-collapse collapse" id="navbar-collapse-4579">
#>       <ul class="nav navbar-nav" data-tabsetid="1081">
#>         <li class="active">
#>           <a href="#tab-1081-1" data-toggle="tab" data-bs-toggle="tab" data-value="A">A</a>
#>         </li>
#>         <li>
#>           <a href="#tab-1081-2" data-toggle="tab" data-bs-toggle="tab" data-value="B">B</a>
#>         </li>
#>       </ul>
#>     </div>
#>   </div>
#> </nav>
#> <div class="container-fluid">
#>   <div class="tab-content" data-tabsetid="1081">
#>     <div class="tab-pane active" data-value="A" id="tab-1081-1">a</div>
#>     <div class="tab-pane" data-value="B" id="tab-1081-2">b</div>
#>   </div>
#> </div>