precompiled_css_path {bslib}	R Documentation

Get the path to a precompiled CSS file

Description

This function is meant for development and debugging purposes. It can be used to test if a precompiled CSS file exists for a given theme object.

Usage

precompiled_css_path(theme = bs_theme())

Arguments

theme
a bs_theme() object.

Value

The path to the precompiled CSS file, if it exists, or NULL, if not.

[Package bslib version 0.5.1 Index]